Simplify actual_merge(), fix the tests.
This now is more like "Top-down implementation using lists".
https://en.wikipedia.org/wiki/Merge_sort#Top-down_implementation_using_lists

Change the names to be left and right to match the psuedo-code and to avoid
tainting the names in this function with context from other functions. They're
just lists, doesn't matter where they came from.

Returning a Result is unnecessary, there's no possibility of an error.
`left/right.pop_front().unwrap()` is safe because we've already checked
they're not empty.

Waiting to pop until after the comparison simplifies the code.

Fix the tests, actual_merge() assumes the lists are already sorted.